[~] Fix time range on first load
