+++ src/hooks/index.js
@@ -11,6 +11,7 @@ export { default as useNetwork } from './useNetwork'; // Graph
 export { default as usePaths } from './usePaths'; // Graph Helpers
 export { default as useQueryVariables } from './useQueryVariables'; // Utils
 export { default as useSelectedNodes } from './useSelectedNodes'; // Graph Helpers
+export { default as useStaticTimeRange } from './useStaticTimeRange'; // Dashboard
 export { default as useStore } from './useStore'; // Utils
-export { default as useTimeRange } from './useTimeRange'; // Dashboard
+export { default as useTimeRange } from './useTimeRange'; // Utils
 export { default as useTypes } from './useTypes'; // Utils
